// src/interfaces/customer.interface.ts
export interface ICustomerAttributes {
  C_Number?: number;
  C_Name?: string;
  C_CoName?: string;
  C_Address?: string;
  C_City?: string;
  C_State?: string;
  C_Zip?: string;
  C_Country?: string;
  AddressType?: any;
  C_Contact1?: string;
  C_Contact2?: string;
  Jurisdiction_State?: number;
  Jurisdiction_County?: number;
  Jurisdiction_City?: number;
  C_Salesman?: number;
  TermsCode?: number;
  C_Phone?: string;
  C_Fax?: string;
  C_SpecialTaxCode?: string;
  C_Class?: number;
  C_StatusCode?: number;
  C_SalesTaxNumber?: string;
  C_CigtLicenseNumber?: string;
  C_StatementAccount?: number;
  C_CaseDiscount?: boolean;
  C_RetailCode?: number;
  C_StatementCode?: number;
  C_Memo?: string;
  C_AuthorizedOnly?: boolean;
  C_ClassOfTrade?: string;
  C_InvoiceFormat?: number;
  C_PricingAccount?: number;
  PriceLevel_Default?: number;
  NetCost_Flag?: boolean;
  Credit_Limit?: number;
  EDI_Format?: number;
  C_CashCustomer?: boolean;
  Finance_Charge?: boolean;
  C_Interest?: number;
  Interest_LastCharged?: Date;
  ExpDate_SalesTax?: Date;
  ExpDate_CigtTax?: Date;
  C_Password?: string;
  C_ChainStore?: string;
  C_Email: string;
  C_Inactive?: boolean;
  Delivery_Charge?: boolean;
  Delivery_Amount?: number;
  C_OtherLicenseNumber?: string;
  ExpDate_OtherTax?: Date;
  C_SalesTaxSelect?: string;
  C_StoreID?: string;
  C_StoreIDQ?: string;
  No_Substitutes?: boolean;
  C_OrderDay?: number;
  C_RetailRounding?: string;
  C_FEIN?: string;
  C_OperationHours1?: number;
  C_OperationHours2?: number;
  Service_Charge?: boolean;
  Other_Amount?: number;
  C_PhoneMobile?: string;
  POS_CashC?: boolean;
  POS_CheckC?: boolean;
  POS_CreditC?: boolean;
  POS_DebitC?: boolean;
  POS_OtherC?: boolean;
  POS_HouseC?: boolean;
  Delivery_ID?: number;
  C_OrderDaySequence?: number;
  C_DateCreated?: Date;
  C_DateCreatedUser?: number;
  C_DateModifed?: Date;
  C_DateModifedUser?: number;
  C_CommissionRate01?: number;
  C_CommissionRate01_PCT?: boolean;
  C_CommissionRate02?: number;
  C_CommissionRate03?: number;
  C_CommissionRate04?: number;
  C_CommissionRate05?: number;
  C_CommissionRate06?: number;
  C_CommissionRate07?: number;
  C_CommissionRate08?: number;
  C_CommissionRate09?: number;
  C_CommissionRate10?: number;
  C_CommissionRate11?: number;
  C_CommissionRate12?: number;
  LastBalance?: number;
  LastInvoiceDate?: Date;
  LastInvoiceNumber?: number;
  LastInvoiceAmount?: number;
  LastPaymentDate?: Date;
  LastPaymentAmount?: number;
  LastPostDatedTotal?: number;
  MSA_AcceptPromo?: boolean;
  WebHistory?: number;
  FTP_Host?: string;
  FTP_User?: string;
  FTP_Password?: string;
  FTP_Directory?: string;
  FTP_Protocol?: number;
  FTP_Mode?: number;
  FTP_FileExt?: string;
  C_OutputFolder?: string;
  C_ReturnStatus?: string;
  C_OtherLicenseNumber2?: string;
  ExpDate_OtherTax2?: Date;
  C_OtherLicenseNumber3?: string;
  ExpDate_OtherTax3?: Date;
  MSA_StateTaxJurisdiction?: string;
  C_PromotionGroup?: string;
  C_Alias?: string;
  Category_Allow01?: boolean;
  Category_Allow02?: boolean;
  Category_Allow03?: boolean;
  Category_Allow04?: boolean;
  Category_Allow05?: boolean;
  Category_Allow06?: boolean;
  Category_Allow07?: boolean;
  Category_Allow08?: boolean;
  Category_Allow09?: boolean;
  Category_Allow10?: boolean;
  Category_Allow11?: boolean;
  Category_Allow12?: boolean;
  emailInvoice?: boolean;
  emailInvoiceEDI?: boolean;
  emailReport?: boolean;
  emailStatement?: boolean;
  emailPromo?: boolean;
  C_AdminOnly?: boolean;
  eCommerce_UpdateTag?: boolean;
  eCommerce_FTP_HostID?: number;
  C_InputFormat?: number;
  C_InputFolder?: string;
  C_StatementFormat?: number;
}